+++ docs/tutorial/chart_layout.js
@@ -25,9 +25,7 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
 				(chart) => {
 					return chart.animate({
 						style: {
-							title: {
-								backgroundColor: '#A0A0A0'
-							},
+							backgroundColor: '#A0A0A0',
 							plot: {
 								backgroundColor: '#D2D2D2'
 							},