+++ src/chart/rendering/drawguides.cpp
@@ -5,6 +5,7 @@
 
 #include "base/geom/line.h"
 #include "base/geom/point.h"
+#include "base/math/fuzzybool.h"
 #include "chart/generator/plot.h" // NOLINT(misc-include-cleaner)
 #include "chart/main/events.h"
 #include "chart/options/channel.h"
@@ -33,14 +34,13 @@ void DrawGuides::draw(bool horizontal)
 	const auto &axis = axises.at(axisId).dimension;
 
 	if (axis.enabled && *guideStyle.lineWidth > 0
-	    && (static_cast<double>(plot->guides.at(axisId).axisGuides)
-	        > 0)) {
+	    && plot->guides.at(axisId).axisGuides != false) {
 		canvas.setLineWidth(*guideStyle.lineWidth);
 
 		for (auto it = axis.begin(); it != axis.end(); ++it) {
-			auto weight = it->second.weight;
-			weight *= static_cast<double>(
-			    plot->guides.at(axisId).axisGuides);
+			auto weight =
+			    Math::FuzzyBool::And<double>(it->second.weight,
+			        plot->guides.at(axisId).axisGuides);
 			if (weight == 0) continue;
 
 			auto next = std::next(it);