+++ src/chart/rendering/drawinterlacing.cpp
@@ -8,7 +8,9 @@
 #include "base/anim/interpolated.h"
 #include "base/geom/point.h"
 #include "base/geom/rect.h"
+#include "base/gfx/colortransform.h"
 #include "base/gfx/font.h"
+#include "base/math/floating.h"
 #include "base/math/range.h"
 #include "base/math/renard.h"
 #include "base/text/smartstring.h"
@@ -78,9 +80,10 @@ void DrawInterlacing::draw(bool horizontal, bool text) const
 	}
 	else {
 		auto highWeight =
-		    Math::Range(stepLow, stepHigh).rescale(step) * enabled;
+		    Math::Range<double>::Raw(stepLow, stepHigh).rescale(step);
 
 		auto lowWeight = (1.0 - highWeight) * enabled;
+		highWeight *= enabled;
 
 		draw(axis.enabled,
 		    horizontal,
@@ -117,28 +120,32 @@ void DrawInterlacing::draw(
 
 	const auto origo = plot->axises.origo();
 
-	if (static_cast<double>(enabled.interlacings || enabled.axisSticks
-	                        || enabled.labels)
-	    > 0) {
-		auto interlaceIntensity =
-		    weight * static_cast<double>(enabled.interlacings);
+	if ((enabled.interlacings || enabled.axisSticks || enabled.labels)
+	    != false) {
+		auto interlaceIntensity = Math::FuzzyBool::And<double>(weight,
+		    enabled.interlacings);
 		auto interlaceColor =
 		    *axisStyle.interlacing.color * interlaceIntensity;
 
 		auto tickIntensity =
-		    weight * static_cast<double>(enabled.axisSticks);
+		    Math::FuzzyBool::And<double>(weight, enabled.axisSticks);
 
-		auto textAlpha = weight * static_cast<double>(enabled.labels);
+		auto textAlpha =
+		    Math::FuzzyBool::And<double>(weight, enabled.labels);
 
-		if (rangeSize <= 0) return;
+		if (std::signbit(rangeSize) != std::signbit(stepSize)
+		    || Math::Floating::is_zero(rangeSize))
+			return;
 
 		auto stripWidth = stepSize / rangeSize;
 
 		auto axisBottom = axis.origo() + stripWidth;
 
-		auto iMin = axisBottom > 0 ? static_cast<int>(
-		                std::floor(-axis.origo() / (2 * stripWidth)))
-		                           : 0;
+		auto iMin =
+		    axisBottom > 0 ? static_cast<int>(
+		        std::floor(-axis.origo() / (2 * stripWidth)))
+		                   : static_cast<int>(
+		                       (axis.range.getMin() - stepSize) / 2);
 
 		if (stripWidth <= 0) return;
 		auto interlaceCount = 0U;
@@ -267,7 +274,7 @@ void DrawInterlacing::drawDataLabel(
 		auto under = labelStyle.position->interpolates()
 		               ? labelStyle.side->get_or_first(index).value
 		                     == Styles::AxisLabel::Side::negative
-		               : labelStyle.side->factor<double>(
+		               : labelStyle.side->factor(
 		                   Styles::AxisLabel::Side::negative);
 
 		auto &&posDir =
@@ -285,8 +292,9 @@ void DrawInterlacing::drawDataLabel(
 		    posDir,
 		    labelStyle,
 		    0,
-		    alpha * position.weight * wUnit.weight,
-		    wUnit.weight,
+		    Gfx::ColorTransform::Opacity(Math::FuzzyBool::And(alpha,
+		        position.weight,
+		        wUnit.weight)),
 		    *rootEvents.draw.plot.axis.label,
 		    Events::Targets::measAxisLabel(str, !horizontal));
 	}